Reference class refactoring
- Enable implicit casting between reference types
- Extract reference classes to new submodule
